+++ src/components/load-more-films-button.js
@@ -1,0 +1,9 @@
+const createShowMoreFilmsTemplate = () => {
+  return (
+    `
+     <button class="films-list__show-more">Show more</button>
+    `
+  );
+};
+
+export {createShowMoreFilmsTemplate};
